feat: add item cards and item grid
This adds an early item card component that simply renders the item's thumb, title and release year.
It also adds an item grid, which is fully virtualized and supports lazy loading (using
react-virtualized). The home page is currently hardcoded to show the contents of the library with
ID=0, as a way to test the components.