build(cra): upload sourcemaps to Sentry
Adds CRACO as a way to modify CRA's config without ejecting. We then add Sentry's webpack plugin to
the config when running a production build. As an easy way to get a version number without including
package.json, this also adds Google's replacement plugin for Semantic Release, which updates a
utility file when building a release.