chore: initial commit
The old repository wasn't great to work with. This moves it from CRA to
a monorepo using NX.
A couple of changes sneaked in during the porting process, such as a
change to the color palette, moving some components to a separate UI
library, adding tests and properly using Storybook, adding a library
creation modal (because I was tired of having to created libraries
manually), support for square cards, changes to the GraphQL layer to go
along with the recent server changes, etc.